Split wires sub-directory into component parts: motion, collection, and analysis
