+++ slac_measurements/wires/analysis/analysis.py
@@ -7,9 +7,9 @@ import warnings
 from typing import Literal
 
 import slac_measurements.beam_profile
-from slac_measurements.wires.coordinates import stage_to_beam, beam_to_stage
-from slac_measurements.wires.jitter_correction import compute_jitter
-from slac_measurements.wires.analysis_results import (
+from slac_measurements.wires.analysis.coordinates import stage_to_beam, beam_to_stage
+from slac_measurements.wires.analysis.jitter_correction import compute_jitter
+from slac_measurements.wires.analysis.results import (
     DetectorFit,
     DetectorProfileMeasurement,
     FitResult,
